% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-biological-replicates.R
\name{add_biological_replicates}
\alias{add_biological_replicates}
\title{Add biological replicate annotations}
\usage{
add_biological_replicates(dir, file, overwrite = FALSE)
}
\arguments{
\item{dir}{Path to screenmill analysis directory}

\item{file}{Name of resulting CSV file.}

\item{overwrite}{A flag indicating whether to overwrite an existing biological replicate
annotation file.}
}
\description{
Creates a file named \code{biological-replicate-annotation.csv} with the
minimal columns necessary to easily annotated biological replicates
for every plate. Currently only supports single timepoint analysis.
}
