% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-genetic-distance.R
\name{add_genetic_distance}
\alias{add_genetic_distance}
\title{Creates a distance column containing the genetic distance between the array gene and the query gene}
\usage{
add_genetic_distance(df)
}
\arguments{
\item{df}{A data.frame with strain_id and query_id columns}
}
\description{
doi:  10.1016/S0076-6879(10)70009-4
}
