% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-plate-normalization.R
\name{apply_plate_normalization}
\alias{apply_plate_normalization}
\title{Apply plate normalization}
\usage{
apply_plate_normalization(sm_data, to)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{to}{A column to apply the normalization to (as a string).}
}
\description{
Adds \code{plate_effect} and \code{size_plate_norm} columns to a screenmill dataset.
}
