% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-spatial-effect-normalization.R
\name{apply_spatial_effect_normalization}
\alias{apply_spatial_effect_normalization}
\title{Apply spatial effect normalization for all plates}
\usage{
apply_spatial_effect_normalization(sm_data, to, death_thresh = 0.25, ...)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{death_thresh}{Death threshold for excluding slow growing strains from model fitting. Defaults to \code{0.25}}

\item{...}{Additional parameters passed to \link[stats:loess]{stats::loess}}
}
\description{
Adds \code{spatial_effect} and \code{size_spatial_norm} columns to a screenmill dataset.
}
\details{
Uses \link[stats:loess]{stats::loess} to fit a two degree polynomial surface of the form $colony ~ row + column$.
}
