% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-dead-strains.R
\name{check_dead_strains}
\alias{check_dead_strains}
\title{Creates 'WARNING_mixed_array.csv' as an easy to read automated report of the problematic plates seemingly mixed or wrong or contaminated.
vsp33 is absent from the array and should always be a blank spot.}
\usage{
check_dead_strains(dir, death_thresh = 10)
}
\arguments{
\item{dir}{A directory containing the screenmill data}

\item{death_tresh}{Death threshold for excluding potential pinning residues. default set to '10', stringent treshold.}
}
\description{
Creates 'WARNING_mixed_array.csv' as an easy to read automated report of the problematic plates seemingly mixed or wrong or contaminated.
vsp33 is absent from the array and should always be a blank spot.
}
