% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-self-crosses.R
\name{check_self_crosses}
\alias{check_self_crosses}
\title{Check for growth of the self-cross. Query strains and array strains K.O. at the same locus should not grow,
because geneX::KanMX and geneX::NatMX cannot be both present in an haploid double mutant.}
\usage{
check_self_crosses(dir, untreated = "CisPt-0uM")
}
\arguments{
\item{dir}{A directory containing the screenmill data}

\item{untreated}{The name of the untreated condition under the column treatment_id (see \link[screenmill:read_screenmill]{screenmill::read_screenmill}) (as a string)}
}
\description{
Adds WARNING_self_crossing as a logical variable in a screenmill dataset and creates 'WARNING_self_crossing.csv'
as an easy to read automated report of the problematic crosses.
}
\details{
This will only be applied to the untreated plates.
}
