% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude-large-colonies.R
\name{exclude_large_colonies}
\alias{exclude_large_colonies}
\title{Exclude large colonies}
\usage{
exclude_large_colonies(sm_data, thresh)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill}).}

\item{thresh}{Threshold for identifying large colonies (recommend using \code{1.5}).}
}
\description{
Exclude large colonies
}
