% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave-one-out-mean.R
\name{leave_one_out_mean}
\alias{leave_one_out_mean}
\title{Generic function that drops abberant values of small distribution (2<n<8) and returns a new median if it significantly reduces the sd}
\usage{
leave_one_out_mean(x)
}
\arguments{
\item{x}{distribution (numeric vector)}
}
\description{
Generic function that drops abberant values of small distribution (2<n<8) and returns a new median if it significantly reduces the sd
}
