% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-batch.R
\name{new_batch}
\alias{new_batch}
\title{Create a new batch report}
\usage{
new_batch(proj_dir = ".", data_dir = "data",
  author = getOption("devtools.name"))
}
\arguments{
\item{proj_dir}{Path to project directory}

\item{data_dir}{Path to data directory}

\item{author}{Name of author}
}
\description{
Writes an Rmarkdown report to the current working directory with a recommended pipeline for
analyzing images, normalizing measurements, and generating quality control tables and figures.
}
