% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-batch-effect.R
\name{normalize_batch_effect}
\alias{normalize_batch_effect}
\title{Batch effect normalization. Relies on the plate controls of each normalised (query and drug effect removed) plates, and correct the bias associated with batches from the same date}
\usage{
normalize_batch_effect(df)
}
\arguments{
\item{df}{A data.frame with plate_id, query_id, date, cisplatin, plate_control_mean}
}
\description{
Batch effect normalization. Relies on the plate controls of each normalised (query and drug effect removed) plates, and correct the bias associated with batches from the same date
}
