% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-spatial-effect.R
\name{normalize_spatial_effect}
\alias{normalize_spatial_effect}
\title{Apply spatial effect normalization for all plates}
\usage{
normalize_spatial_effect(sm_data, of, death_thresh, prefix, ...)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{of}{A column to apply the normalization to (as a string).}

\item{death_thresh}{Death threshold for excluding slow growing strains from model fitting.}

\item{prefix}{A prefix appended to \code{plate_effect} and \code{plate_norm} column names.}

\item{...}{Additional parameters passed to \link[stats:loess]{stats::loess}}
}
\description{
Adds \code{spatial_effect} and \code{size_spatial_norm} columns to a screenmill dataset.
}
\details{
Uses \link[stats:loess]{stats::loess} to fit a two degree polynomial surface of the form $colony ~ row + column$.
}
