% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sm-data.R
\name{plot_sm_data}
\alias{plot_sm_data}
\title{Creates different figures with ggplot2 to visualize the data and the quality controls steps.}
\usage{
plot_sm_data(sm_data, normalized = TRUE, path)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{normalized}{Logical. Has the data been spatially and plate normalized ? default is TRUE (see\link[rothSGA:apply_spatial_effect_normalization]{rothSGA::apply_spatial_effect_normalization} and \link[rothSGA:apply_plate_normalization]{rothSGA::apply_plate_normalization})}

\item{path}{where the figures will be generated. A figures directory will be created.}
}
\description{
Creates different figures with ggplot2 to visualize the data and the quality controls steps.
}
