% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothSGA-package.R
\name{rothSGA}
\alias{rothSGA}
\title{SGA analysis tools for the Rothstein Lab.}
\description{
Tools in this package are provided to standardize common steps in the
quantification of genetic interactions derived from Synthetic Genetic Array
(\href{https://en.wikipedia.org/wiki/Synthetic_genetic_array}{SGA}) colony growth
experiments.
}
\details{
Groups of SGA experiments performed in parallel at the same time
are considered part of the same "Batch". We recommend processing each
batch independently using approximately the following workflow:
\enumerate{
\item Record colony growth (ideal incubation time will depend on several factors
including: strain background, media, temperature, pinning procedure). We
recommend using a flatbed scanner with a transparency lid (e.g. the
\href{https://www.amazon.com/Microtek-ScanMaker-Scanner-Transparent-Adapter/dp/B000068CNE}{Micotek 9800XL}
can scan 9 standard rectangular plates.)
\item Annotate plates in images with \link[screenmill:annotate]{screenmill::annotate} which integrates
with the Rothstein Lab's annotation database \link[rothfreezer:rothfreezer]{rothfreezer::rothfreezer}
\item Locate colony grid and determine plate cropping coordinates with
\link[screenmill:calibrate]{screenmill::calibrate}
\item Review plates to exclude regions with contamination or failed pinnings
with \link[screenmill:review]{screenmill::review}
\item Measure colonies with \link[screenmill:measure]{screenmill::measure}
\item Normalize spatial and plate effects with
\link[rothSGA:normalize_spatial_effect]{rothSGA::normalize_spatial_effect} and \link[rothSGA:normalize_plate_effect]{rothSGA::normalize_plate_effect}
\item Run quality control checks with e.g. \link[rothSGA:check_dead_strains]{rothSGA::check_dead_strains} and
\link[rothSGA:check_self_crosses]{rothSGA::check_self_crosses}
}

This pipeline can be quickly generated using \link[rothSGA:new_batch_report]{rothSGA::new_batch_report}
which generates an Rmarkdown report from a standard template.
}
