% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-normalized-data.R
\name{summarize_normalized_data}
\alias{summarize_normalized_data}
\title{Aggregates and summarizes technical replicates on the plate}
\usage{
summarize_normalized_data(sm_data, to = size_plate_norm)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{to}{A column to apply the stats. default is size_plate_norm (see\link[rothSGA:apply_plate_normalization]{rothSGA::apply_plate_normalization}) (as a string).}
}
\description{
Adds mean_size_norm , median_size_norm, and sd_size_norm to a screenmill dataset.
}
