% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-plate-effect.R
\name{normalize_plate_effect}
\alias{normalize_plate_effect}
\title{Apply plate normalization}
\usage{
normalize_plate_effect(sm_data, of, prefix)
}
\arguments{
\item{sm_data}{A screenmill dataset (see \link[screenmill:read_screenmill]{screenmill::read_screenmill})}

\item{of}{A column to apply the normalization to (as a string).}

\item{prefix}{A prefix appended to \code{plate_effect} and \code{plate_norm} column names.}
}
\description{
Adds \code{plate_effect} and \code{size_plate_norm} columns to a screenmill dataset.
}
